replace old tabs with new tabs
