--- lib/components/Tabs/Tab.mdx
@@ -1,27 +1,0 @@
-import * as stories from "./Tab.stories";
-import { Meta, Story, Controls, Canvas } from "@storybook/addon-docs";
-import { TabsContainer, Tab } from ".";
-
-<Meta of={stories} />
-
-[Default tabs](#tabs) | [Properties](#properties)
-
-# Tabs
-
-Tabs are intended to provide a way to navigate between distinct sections within the layout without leaving the page or triggering a page refresh.
-
-## Active tab
-
-<Canvas of={stories.activeTab} />
-
-## With notifications tab
-
-<Canvas of={stories.withNotificationsTab} />
-
-## With popover tab
-
-<Canvas of={stories.withPopoverTab} />
-
-## Properties
-
-<Controls />
