+++ lib/components/Tabs/Tabs.mdx
@@ -1,14 +1,16 @@
-import { Meta, Story, Controls, Canvas } from "@storybook/addon-docs";
-import { TabsContainer, Tab } from ".";
-import * as stories from "./TabsContainer.stories";
+import { Story, Controls, Meta, Canvas } from "@storybook/addon-docs";
+import Tabs from ".";
+import * as stories from "./Tabs.stories";
 
 <Meta of={stories} />
 
-[Default tabs](#tabs) | [Properties](#properties)
-
 # Tabs
 
+[Default Tabs](#default-tabs) | [Properties](#properties)
+
 Tabs are intended to provide a way to navigate between distinct sections within the layout without leaving the page or triggering a page refresh.
+
+## Default Tabs
 
 <Canvas of={stories.defaultTabs} />
