+++ lib/components/Tabs/Tabs.stories.js
@@ -1,10 +1,10 @@
 import React from "react";
-import TabsV2 from ".";
+import Tabs from ".";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 import Box from "../Box";
 
 export default {
-  title: "Components/TabsV2",
+  title: "Components/Tabs",
   decorators: [
     (storyFn) => (
       <Box bg="greyLightest" p="xl" pb="300px">
@@ -12,7 +12,7 @@ export default {
       </Box>
     )
   ],
-  component: TabsV2
+  component: Tabs
 };
 
 const tabsList = [
@@ -54,9 +54,9 @@ const tabsList = [
   }
 ];
 
-export const defaultTabsV2 = () => (
+export const defaultTabs = () => (
   <BrowserRouter>
-    <TabsV2 tabsList={tabsList} />
+    <Tabs tabsList={tabsList} />
     <Switch>
       {tabsList.map((tab) => (
         <Route key={tab.path} path={tab.path}>
@@ -68,4 +68,4 @@ export const defaultTabsV2 = () => (
     </Switch>
   </BrowserRouter>
 );
-defaultTabsV2.storyName = "Default";
+defaultTabs.storyName = "Default";
