+++ lib/components/Tabs/index.js
@@ -1,145 +1,271 @@
-import React from "react";
-import PropTypes from "prop-types";
-import styled, { ThemeProvider } from "styled-components";
+import React, { useState, useRef, useEffect, useCallback } from "react";
+import styled, { css } from "styled-components";
+import { NavLink, useLocation } from "react-router-dom";
+import { isEqual } from "lodash";
+import ActionsMenu, { ActionsMenuItem } from "../ActionsMenu";
+import Icon from "../Icon";
+import FlexItem from "../Flex";
 import { themeGet } from "@styled-system/theme-get";
-import { css } from "@styled-system/css";
-import { space, color } from "styled-system";
+import PropTypes from "prop-types";
 
-export const TabsContainerItem = styled("div")(
-  css({
-    width: "100%",
-    display: "flex",
-    alignItems: "center",
-    flexWrap: "wrap",
-    bg: "transparent"
-  }),
-  space
-);
+const TabsContainer = styled.div`
+  position: relative;
+`;
 
-export const TabItem = styled("div")(
-  (props) =>
-    css({
-      marginRight: 3,
-      a: {
-        borderRadius: themeGet("radii.2")(props),
-        bg: themeGet("colors.greyLighter")(props),
-        transition: themeGet("transition.transitionDefault")(props),
-        px: 4,
-        py: 3,
-        fontSize: themeGet("fontSizes.1")(props),
-        fontWeight: themeGet("fontWeights.2")(props),
-        color: themeGet("colors.greyDarker")(props),
-        display: "flex",
-        alignItems: "center",
-        position: "relative",
-        whiteSpace: "nowrap",
-        textDecoration: "none",
-        textAlign: "center",
-        textTransform: "uppercase",
-        cursor: "pointer",
-        "&:hover": {
-          bg: themeGet("colors.greyLight")(props),
-          color: themeGet("colors.greyDarker")(props),
-          outline: "0"
-        },
-        "&:focus": {
-          color: themeGet("colors.greyDarker")(props),
-          outline: "0",
-          boxShadow:
-            themeGet("shadows.thinOutline")(props) +
-            " " +
-            themeGet("colors.grey")(props)
-        },
-        button: {
-          marginLeft: 2
-        }
-      },
-      "&.active": {
-        a: {
-          color: themeGet("colors.primary")(props),
-          bg: themeGet("colors.white")(props),
-          "&:hover": {
-            cursor: "default"
+const TabWrapper = styled.div`
+  position: relative;
+`;
+
+const VisibleTabs = styled.div`
+  flex-shrink: 1;
+  display: flex;
+  align-items: center;
+  overflow: hidden;
+`;
+
+const activeTabStyle = css`
+  background-color: ${themeGet("colors.white")};
+  color: ${themeGet("colors.primary")};
+  cursor: default;
+  &:hover {
+    background-color: ${themeGet("colors.white")};
+    color: ${themeGet("colors.primary")};
+  }
+  &:focus {
+    color: ${themeGet("colors.primary")};
+  }
+`;
+
+const Tab = styled(NavLink)`
+  width: ${({ fullWidth }) => (fullWidth ? "100%" : "fit-content")};
+  display: block;
+  border-radius: ${themeGet("radii.2")};
+  transition: background 200ms ease-in-out, color 200ms ease-in-out;
+  padding: ${themeGet("space.3")} ${themeGet("space.4")};
+  font-size: ${themeGet("fontSizes.1")};
+  font-weight: ${themeGet("fontWeights.2")};
+  position: relative;
+  white-space: nowrap;
+  text-decoration: none;
+  text-align: center;
+  text-transform: uppercase;
+  background-color: ${themeGet("colors.greyLighter")};
+  color: ${themeGet("colors.greyDarker")};
+  cursor: pointer;
+
+  ${({ tabInShowMore }) =>
+    tabInShowMore
+      ? css`
+          position: absolute;
+          visibility: hidden;
+        `
+      : ""}
+
+  &:hover {
+    outline: 0;
+    background-color: ${themeGet("colors.greyLight")};
+    color: ${themeGet("colors.greyDarker")};
+  }
+  &:focus {
+    outline: 0;
+    color: ${themeGet("colors.greyDarker")};
+    box-shadow: inset ${themeGet("shadows.thinOutline")}
+      ${themeGet("colors.grey")};
+  }
+
+  &.active {
+    ${activeTabStyle}
+  }
+`;
+
+const ShowMoreButton = styled.button`
+  appearance: none;
+  border: none;
+  font-family: ${themeGet("fonts.main")};
+  font-size: ${themeGet("fontSizes.1")};
+  font-weight: ${themeGet("fontWeights.2")};
+  border-radius: ${themeGet("radii.2")};
+  background-color: ${themeGet("colors.greyLighter")};
+  transition: ${themeGet("transition.transitionDefault")};
+  padding: ${themeGet("space.3")} ${themeGet("space.4")};
+  color: ${themeGet("colors.greyDarker")};
+  display: ${({ showMoreVisible }) => (showMoreVisible ? "flex" : "none")};
+  align-items: center;
+  &:hover {
+    background-color: ${themeGet("colors.greyLight")};
+    color: ${themeGet("colors.greyDarker")};
+    outline: 0;
+  }
+  &:focus {
+    color: ${themeGet("colors.greyDarker")};
+    outline: 0;
+    box-shadow: inset ${themeGet("shadows.thinOutline")}
+      ${themeGet("colors.grey")};
+  }
+  &.hasActive {
+    ${activeTabStyle}
+  }
+`;
+
+const ShowMoreTabs = styled.div`
+  border-radius: ${themeGet("radii.2")};
+  background-color: ${themeGet("colors.white")};
+  min-width: 84px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 6px;
+  padding: 6px;
+  box-shadow: inset ${themeGet("shadows.thinOutline")}
+      ${themeGet("colors.greyLighter")},
+    ${themeGet("shadows.boxDefault")};
+`;
+
+const tabsGap = 6;
+
+const Tabs = ({ tabsList }) => {
+  const containerRef = useRef(null);
+
+  const showMoreButtonRef = useRef();
+
+  const containerVisibleWidth = useRef(0);
+
+  const [showMoreTabs, setShowMoreTabs] = useState([]);
+
+  const calculateVisibility = useCallback(
+    (actionElements) => {
+      const showMoreButtonWidth = showMoreButtonRef.current
+        ? showMoreButtonRef.current.offsetWidth
+        : 0;
+      // as we loop through the tabs, we need to calculate the width of the visible tabs.
+      let calculatedWidth = showMoreTabs.length ? showMoreButtonWidth : 0;
+
+      // variable for the list of hidden tabs which will be put to react state
+      const newShowMoreTabs = [];
+
+      [...actionElements]
+        .filter((el) => el.tagName === "A")
+        .forEach((actionEl, i) => {
+          // visibleElementsWidth will be increased by
+          // the corresponding width of the element + gapWidth
+          calculatedWidth += actionEl.offsetWidth + tabsGap;
+
+          // compare computed widths and push into newShowMoreTabs if current tab width is bigger than container width
+          if (calculatedWidth >= containerVisibleWidth.current) {
+            newShowMoreTabs.push(i);
           }
-        }
-      },
-      "&.notification": {
-        a: {
-          "&::after": {
-            position: "absolute",
-            top: "calc(8px * -1)",
-            right: "calc(4px * -2)",
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            width: "20px",
-            height: "20px",
-            borderRadius: "50%",
-            fontSize: themeGet("fontSizes.0")(props),
-            fontWeight: themeGet("fontWeights.2")(props),
-            content: `"${props.notification}"`,
-            bg: themeGet("colors.danger")(props),
-            color: themeGet("colors.white")(props),
-            zIndex: 4
-          }
+        });
+
+      if (!isEqual(showMoreTabs, newShowMoreTabs)) {
+        // update React state with the list of hidden tabs
+        setShowMoreTabs(newShowMoreTabs);
+      }
+    },
+    [showMoreTabs]
+  );
+
+  useEffect(() => {
+    const actionElements = containerRef.current?.children || [];
+
+    const resizeObserver = new ResizeObserver((entries) => {
+      for (const entry of entries) {
+        if (entry.contentBoxSize) {
+          const contentBoxSize = entry.contentBoxSize[0];
+
+          // Math.ceil is necessary to round up and return
+          // the smallest integer for the size of observed element
+          containerVisibleWidth.current = Math.ceil(contentBoxSize.inlineSize);
+
+          // invoke the functions which calculates tabs visibility
+          // and sets data to the list of hidden tabs
+          calculateVisibility(actionElements);
         }
       }
-    }),
-  space,
-  color
-);
+    });
 
-export const Tab = ({ theme, children, active, notification, ...props }) => {
-  const component = (
-    <TabItem
-      active={active}
-      notification={notification}
-      role="presentation"
-      className={`${active ? "active" : ""} ${
-        notification ? "notification notification-" + notification : ""
-      }`}
-      {...props}
-    >
-      {React.Children.map(children, (child) =>
-        React.cloneElement(child, { role: "tab", "aria-selected": `${active}` })
-      )}
-    </TabItem>
+    // adding ResizeObserver to the observed container
+    resizeObserver.observe(containerRef.current);
+  }, [calculateVisibility]);
+
+  const visibleTabsList = tabsList
+    .map((tab) => ({ ...tab, isVisible: tab.isVisible ?? true }))
+    .filter((tab) => tab.isVisible);
+  const showMoreTabsList = visibleTabsList.filter((tab, i) =>
+    showMoreTabs.includes(i)
   );
-  return theme ? (
-    <ThemeProvider theme={theme}>{component}</ThemeProvider>
-  ) : (
-    component
+  const location = useLocation();
+  const showMoreItemActive = showMoreTabsList.find((action) =>
+    location.pathname.endsWith(action.path)
+  );
+
+  return (
+    <TabsContainer>
+      <TabWrapper>
+        <VisibleTabs style={{ gap: tabsGap }} ref={containerRef}>
+          {visibleTabsList.map((tab, i) => (
+            <Tab
+              className={tab.className}
+              id={tab.id}
+              data-testid={tab["data-testid"]}
+              key={tab.path}
+              tabInShowMore={showMoreTabs.includes(i)}
+              to={tab.path}
+            >
+              {tab.label}
+            </Tab>
+          ))}
+
+          <ActionsMenu
+            direction="bottom-end"
+            customTriggerComponent={
+              <ShowMoreButton
+                ref={showMoreButtonRef}
+                showMoreVisible={showMoreTabsList.length}
+                className={showMoreItemActive && "hasActive"}
+              >
+                <FlexItem flex="0 0 auto">More</FlexItem>
+                <FlexItem flex="0 0 auto">
+                  <Icon
+                    icon={["fas", "chevron-down"]}
+                    title="Down"
+                    ml="s"
+                    size="sm"
+                  />
+                </FlexItem>
+              </ShowMoreButton>
+            }
+            closeOnClick
+          >
+            <ShowMoreTabs>
+              {showMoreTabsList.map((tab) => (
+                <ActionsMenuItem
+                  className={tab.className}
+                  id={tab.id}
+                  data-testid={tab["data-testid"]}
+                  key={tab.path}
+                  Component={Tab}
+                  fullWidth
+                  to={tab.path}
+                >
+                  {tab.label}
+                </ActionsMenuItem>
+              ))}
+            </ShowMoreTabs>
+          </ActionsMenu>
+        </VisibleTabs>
+      </TabWrapper>
+    </TabsContainer>
   );
 };
 
-Tab.propTypes = {
-  /** Specifies whether the tab is the active tab */
-  active: PropTypes.bool,
-  /** Specifies any notifications attached to the Tab */
-  notification: PropTypes.string,
-  /** Specifies the colour theme */
-  theme: PropTypes.object,
-  /** The content of the Tab is passed as a child. */
-  children: PropTypes.node
+Tabs.propTypes = {
+  tabsList: PropTypes.arrayOf(
+    PropTypes.objectOf({
+      label: PropTypes.string,
+      path: PropTypes.string,
+      isVisible: PropTypes.bool
+    })
+  )
 };
 
-export const TabsContainer = ({ theme, children, ...props }) => {
-  const component = (
-    <TabsContainerItem role="tablist" {...props}>
-      {children}
-    </TabsContainerItem>
-  );
-
-  return theme ? (
-    <ThemeProvider theme={theme}>{component}</ThemeProvider>
-  ) : (
-    component
-  );
-};
-
-TabsContainer.propTypes = {
-  /** The contents of the TabsContainer are passed as a child. */
-  children: PropTypes.node,
-  /** Specifies the colour theme of the container */
-  theme: PropTypes.object
-};
+export default Tabs;
