--- lib/components/TabsV2/TabsV2.mdx
@@ -1,19 +1,0 @@
-import { Story, Controls, Meta, Canvas } from "@storybook/addon-docs";
-import Tabs from ".";
-import * as stories from "./TabsV2.stories";
-
-<Meta of={stories} />
-
-# Tabs
-
-[Default Tabs](#default-tabs) | [Properties](#properties)
-
-Tabs are intended to provide a way to navigate between distinct sections within the layout without leaving the page or triggering a page refresh.
-
-## Default Tabs
-
-<Canvas of={stories.defaultTabsV2} />
-
-## Properties
-
-<Controls />
