+++ lib/index.js
@@ -56,8 +56,7 @@ export { default as Spacer } from "./components/Spacer";
 export { default as StatusDot } from "./components/StatusDot";
 export { default as StyledLink, styleLink } from "./components/StyledLink";
 export { default as Table } from "./components/Table";
-export { Tab, TabsContainer } from "./components/Tabs";
-export { default as TabsV2 } from "./components/TabsV2";
+export { default as Tabs } from "./components/Tabs";
 export { default as Tag } from "./components/Tag";
 export { default as TextInput } from "./components/TextInput";
 export { default as TextArea } from "./components/TextArea";
